Make PixelImage global
